+++ oap-server/server-query-plugin/traceql-plugin/src/main/java/org/apache/skywalking/oap/query/traceql/handler/SkyWalkingTraceQLApiHandler.java
@@ -1,0 +1,110 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package org.apache.skywalking.oap.query.traceql.handler;
+
+import com.linecorp.armeria.common.HttpResponse;
+import java.io.IOException;
+import java.util.Optional;
+import org.apache.commons.codec.DecoderException;
+import org.apache.skywalking.oap.server.core.CoreModule;
+import org.apache.skywalking.oap.server.core.query.TraceQueryService;
+import org.apache.skywalking.oap.server.library.module.ModuleManager;
+
+/**
+ * SkyWalking-native implementation of TraceQL API Handler.
+ */
+public class SkyWalkingTraceQLApiHandler extends TraceQLApiHandler {
+    private final TraceQueryService traceQueryService;
+    private final ModuleManager moduleManager;
+
+    public SkyWalkingTraceQLApiHandler(ModuleManager moduleManager) {
+        super();
+        this.moduleManager = moduleManager;
+        this.traceQueryService = moduleManager.find(CoreModule.NAME)
+                                              .provider()
+                                              .getService(TraceQueryService.class);
+    }
+
+    @Override
+    protected HttpResponse queryTraceImpl(String traceId, Optional<String> accept) throws IOException, DecoderException {
+        // TODO: Implement SkyWalking native trace query
+        // 1. Query trace from TraceQueryService
+        // 2. Convert SkyWalking trace format to OTLP format
+        // 3. Return based on Accept header (JSON or Protobuf)
+        return HttpResponse.ofJson("Not implemented yet");
+    }
+
+    @Override
+    protected HttpResponse searchImpl(Optional<String> query,
+                                       Optional<String> tags,
+                                       Optional<String> minDuration,
+                                       Optional<String> maxDuration,
+                                       Optional<Integer> limit,
+                                       Optional<Long> start,
+                                       Optional<Long> end,
+                                       Optional<Integer> spss) throws IOException {
+        // TODO: Implement SkyWalking native trace search
+        // 1. Parse TraceQL query parameters
+        // 2. Build SkyWalking query conditions
+        // 3. Query traces from TraceQueryService
+        // 4. Convert to SearchResponse format
+        return HttpResponse.ofJson("Not implemented yet");
+    }
+
+    @Override
+    protected HttpResponse searchTagsImpl(Optional<String> scope,
+                                          Optional<Integer> limit,
+                                          Optional<Long> start,
+                                          Optional<Long> end) throws IOException {
+        // TODO: Implement SkyWalking tag search
+        // 1. Query available tags from SkyWalking
+        // 2. Filter by scope if provided
+        // 3. Return TagNamesV2Response
+        return HttpResponse.ofJson("Not implemented yet");
+    }
+
+    @Override
+    protected HttpResponse searchTagsV2Impl(Optional<String> q,
+                                            Optional<String> scope,
+                                            Optional<Integer> limit,
+                                            Optional<Long> start,
+                                            Optional<Long> end) throws IOException {
+        // TODO: Implement SkyWalking tag search v2
+        // 1. Parse TraceQL query if provided
+        // 2. Query available tags from SkyWalking
+        // 3. Filter by scope if provided
+        // 4. Return TagNamesV2Response with scopes
+        return HttpResponse.ofJson("Not implemented yet");
+    }
+
+    @Override
+    protected HttpResponse searchTagValuesImpl(String tagName,
+                                                Optional<String> query,
+                                                Optional<Integer> limit,
+                                                Optional<Long> start,
+                                                Optional<Long> end) throws IOException {
+        // TODO: Implement SkyWalking tag value search
+        // 1. Parse TraceQL query if provided
+        // 2. Query tag values from SkyWalking based on tagName
+        // 3. Handle special tags like service.name, span.name, etc.
+        // 4. Return TagValuesResponse
+        return HttpResponse.ofJson("Not implemented yet");
+    }
+}
+