+++ docs/en/changes/changes.md
@@ -142,6 +142,7 @@
 * Bump up netty to 4.2.10.Final.
 * Bump up log4j to 2.25.3 and jackson to 2.18.5.
 * Replace PowerMock Whitebox with standard Java Reflection in `server-library`, `server-core`, and `server-configuration` to support JDK 25+.
+* Support TraceQL and Tempo API for Zipkin trace query.
 
 #### UI
 * Fix the missing icon in new native trace view.